added single summary box
added single summary box option. This will also be stored in your config.

Note that the default is a single summary window. If you don't have it set; this will be chosen for you.

The full update will be pushed to ankiweb once this has been more thouroughly tested (probably done the next time I work on the project).